Fixed the size of the Hackhub logo
I removed most of the border from the Hackhub logo png file. Hackhub is on the third row of the sponsors section and I was wondering if there are going to be more sponsors since if there are we should probably leave it as is, but if there isn't then perhaps expanding everything else may look a lot better.